Add polar grid example notebook (and fix some things)

+++ zea/simulator.py
@@ -106,11 +106,9 @@ def simulate_rf(
             * sound_speed
         )
 
-    n_ax_rounded = _round_up_to_power_of_two(int(n_ax))
+    n_ax_rounded = _round_up_to_power_of_two(int(n_ax)).astype("float32")
 
-    freqs = (
-        ops.cast(ops.arange(n_ax_rounded // 2 + 1) / n_ax_rounded, "float32") * sampling_frequency
-    )
+    freqs = ops.arange(n_ax_rounded // 2 + 1, dtype="float32") / n_ax_rounded * sampling_frequency
 
     waveform_spectrum = pulse_spectrum_fn(freqs)
     parts = []
